BUG: Debugging cuda build, WIP

+++ examples/mesh_form/include/mesh_io.h
@@ -724,8 +724,8 @@ void write_vtu(swage::Mesh& mesh,
     
     // Point vector variables
     for (int var = 0; var < num_point_vec_vars; var++) {
-        fprintf(vtu_file, "        <DataArray type=\"Float32\" Name=\"%s\" NumberOfComponents=\"%d\" format=\"ascii\">\n", 
-                point_vec_var_names[var], num_dims);
+        fprintf(vtu_file, "        <DataArray type=\"Float32\" Name=\"%s\" NumberOfComponents=\"%d\" format=\"ascii\">\n",
+                point_vec_var_names[var], (int)num_dims);
         for (size_t node_gid = 0; node_gid < num_nodes; node_gid++) {
             for (int dim = 0; dim < num_dims; dim++) {
                 fprintf(vtu_file, "          %f\n", vec_fields(node_gid, var, dim));
@@ -750,8 +750,8 @@ void write_vtu(swage::Mesh& mesh,
     
     // Cell vector variables
     for (int var = 0; var < num_cell_vec_vars; var++) {
-        fprintf(vtu_file, "        <DataArray type=\"Float32\" Name=\"%s\" NumberOfComponents=\"%d\" format=\"ascii\">\n", 
-                cell_vec_var_names[var], num_dims);
+        fprintf(vtu_file, "        <DataArray type=\"Float32\" Name=\"%s\" NumberOfComponents=\"%d\" format=\"ascii\">\n",
+                cell_vec_var_names[var], (int)num_dims);
         for (size_t elem_gid = 0; elem_gid < num_elems; elem_gid++) {
             // TODO: Populate cell vector field data from appropriate source
             for (int dim = 0; dim < num_dims; dim++) {
@@ -818,7 +818,7 @@ void write_vtu(swage::Mesh& mesh,
         fprintf(pvtu_file, "    <PPointData>\n");
         for (int var = 0; var < num_point_vec_vars; var++) {
             fprintf(pvtu_file, "      <PDataArray type=\"Float32\" Name=\"%s\" NumberOfComponents=\"%d\"/>\n",
-                    point_vec_var_names[var], num_dims);
+                    point_vec_var_names[var], (int)num_dims);
         }
         for (int var = 0; var < num_point_scalar_vars; var++) {
             fprintf(pvtu_file, "      <PDataArray type=\"Float32\" Name=\"%s\"/>\n",
@@ -830,7 +830,7 @@ void write_vtu(swage::Mesh& mesh,
         fprintf(pvtu_file, "    <PCellData>\n");
         for (int var = 0; var < num_cell_vec_vars; var++) {
             fprintf(pvtu_file, "      <PDataArray type=\"Float32\" Name=\"%s\" NumberOfComponents=\"%d\"/>\n",
-                    cell_vec_var_names[var], num_dims);
+                    cell_vec_var_names[var], (int)num_dims);
         }
         for (int var = 0; var < num_cell_scalar_vars; var++) {
             fprintf(pvtu_file, "      <PDataArray type=\"Float32\" Name=\"%s\"/>\n",
