+++ examples/mesh_form/src/mesh_form.cpp
@@ -112,7 +112,7 @@ int main(int argc, char** argv) {
     binary_stl_reader("/home/jacobmoore/Desktop/repos/ELEMENTS/examples/mesh_form/stl_files/Sphere_superfine.stl", stl_data);
 
     std::cout << "Number of facets: " << stl_data.num_facets << std::endl;
-    printf("Number of boundary nodes: %d\n", mesh.num_bdy_nodes);
+    printf("Number of boundary nodes: %d\n", (int)mesh.num_bdy_nodes);
     
     stl_data.buildAABBTree();
     stl_data.verifyAABBTree();
